Add additional gems for development and setup secret keybase
Added gems 'figaro', 'rspec-rails', 'better_errors' and 'binding_of_caller' to the Gemfile to improve the app in the development and testing phases. The secret key base is now set up in the development environment using figaro, which is added to the gitignore file to keep the sensitive information secure. Also, the 'config/master.key' is included in the .gitignore file. The Gemfile.lock has been updated accordingly.