"Implement flash message animation"
Added an animation effect to the flash message for better user experience. Flash messages now slide from hidden to centered position on screen, remain there for a while and then slide out, instead of just fading out as before. This was done by modifying transform styles in the flash controller and adjusting CSS classes in the application layout view file.