Configure Redis URL via environment variable and Docker setup changes
The Redis URL used in the application has been updated to be configured via an environment variable. This change enables easier customization of the Redis server location, especially when deploying in different environments.

Also, the description of services in the docker-compose.yml file has been modified. The db and redis service configurations are now at the top of the services list. Healthcheck for the database and a condition to ensure that the Redis service has started before the web service starts have been added. Introduction of nginx service is also included in this commit.

This changes make it easier to manage services in a Docker environment efficiently.