Add container names and adminer service to docker-compose.yml
Added specified container names to each service in the docker-compose file for clear identification purposes. Also, included 'adminer' service to the compose file for managing administration tasks on database easily. This service depends on 'db' service and maps the adminer service port 8080 to 9090 on the host. Updated the 'server_name' directive in the nginx configuration file 'default.conf' to point to 'litcode.es' and 'www.litcode.es'.