Enable and enhance server configuration
The server configuration in the `default.conf` file has been updated and enhanced. The configurations that were previously commented out have now been enabled to become active. This includes the addition of a block that listens to port 80 and redirects requests to HTTPS, for added security.

The CORS headers, now also added to the block of code that listens to port 80, were needed to handle Cross-Origin Requests, an essential part for any server that has to respond to requests from different origins. Most importantly, the X-Forwarded-Port header has been set to include the server port in the proxied request, which will allow the backend application to understand the original port that the user tried to access, improving the middleware handling of requests.